// cli/src/types/openapi-schemas.ts
export const TagItem = z.object({ name: z.string() })
// ...
export const AppDescribeInfo = z
  .object({
    author: z.union([z.string(), z.null()]).optional(),
    description: z.union([z.string(), z.null()]).optional(),
    id: z.string(),
    is_agent: z.boolean().optional().default(false),
    mode: z.string(),
    name: z.string(),
    service_api_enabled: z.boolean(),
    tags: z.array(TagItem).optional().default([]),
    updated_at: z.union([z.string(), z.null()]).optional(),
  })
// ...
export type TagItemType = z.infer<typeof TagItem>
export type AppListRowType = z.infer<typeof AppListRow>
export type AppListResponseType = z.infer<typeof AppListResponse>
export type AppDescribeInfoType = z.infer<typeof AppDescribeInfo>